Extending simulator with multi-node feature + testing against known good values
