+++ common/cosmwasm-smart-contracts/mixnet-contract/src/helpers.rs
@@ -3,8 +3,8 @@
 
 use cosmwasm_std::Decimal;
 
-pub fn compare_decimals(a: Decimal, b: Decimal) {
-    let epsilon = Decimal::from_ratio(1u128, 100_000_000u128);
+pub fn compare_decimals(a: Decimal, b: Decimal, epsilon: Option<Decimal>) {
+    let epsilon = epsilon.unwrap_or_else(|| Decimal::from_ratio(1u128, 100_000_000u128));
     if a > b {
         assert!(a - b < epsilon, "{} != {}", a, b)
     } else {
