+++ common/cosmwasm-smart-contracts/mixnet-contract/src/rewarding/mod.rs
@@ -6,7 +6,6 @@ use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 pub mod helpers;
-pub mod multi_node_simulator;
 pub mod simulator;
 
 #[cfg_attr(feature = "generate-ts", derive(ts_rs::TS))]
